requirements.txt
streamlit
pandas
chromadb
google-genai
google-cloud-aiplatform
google-auth
vertexai
openpyxl
plotly
scikit-learn
matplotlib
wordcloud
numpy
google-cloud-bigquery
